Split modules for model and managers
